Dodawanie załączników do nowej publikacji

+++ app/react-publications/src/components/Publication.js
@@ -11,16 +11,21 @@ const Publication = ({ createMode, publication, history, refreshPublications, gl
       pageCount: null,
       publicationYear: null,
       attachments: [],
-      shareList: []
+      shareList: [],
+      links: []
     };
   }
 
   const orginalPublication = publication;
-  console.log(orginalPublication);
 
   const [editMode, setEditMode] = useState(createMode ? true : false);
   const [currentPublication, setCurrentPublication] = useState(orginalPublication);
+  const [toAttachList, setToAttachList] = useState([]);
   const inputClass = 'form-control' + (!editMode ? ' form-control-plaintext' : '');
+  const attachLink = currentPublication.links.find((l) => l.rel === "attachFile");
+
+  console.log(currentPublication);
+
 
   const validatePublication = (publication) => {
     return (
@@ -64,12 +69,19 @@ const Publication = ({ createMode, publication, history, refreshPublications, gl
     let url = baseUrl + globalState.actions['publication.create'].href;
     url = url.replace('{user}', globalState.login);
 
+    const publication = { ...currentPublication };
+    if (editMode) // Clear attachment id
+      publication.attachments = publication.attachments.map(a => ({
+        userName: a.userName,
+        fileName: a.fileName
+      }));
+
     const response = await fetch(url, {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
       },
-      body: JSON.stringify(currentPublication),
+      body: JSON.stringify(publication),
     });
 
     if (!response.ok) {
@@ -82,21 +94,65 @@ const Publication = ({ createMode, publication, history, refreshPublications, gl
   };
 
   const detachFile = async (id) => {
-    const attachment = currentPublication.attachments.find((a) => a.id === id);
-    const url = attachment.links.find((l) => l.rel === 'detach').href;
+    if (!createMode) { // In create mode just update model
+      const attachment = currentPublication.attachments.find((a) => a.id === id);
+      const url = attachment.links.find((l) => l.rel === 'detach').href;
 
-    const response = await fetch(url, { method: 'DELETE' });
-    if (!response.ok) {
-      alert('Nie udało się odłączyć pliku');
-      return;
+      const response = await fetch(url, { method: 'DELETE' });
+      if (!response.ok) {
+        alert('Nie udało się odłączyć pliku');
+        return;
+      }
+      refreshPublications();
     }
 
     setCurrentPublication({
       ...currentPublication,
       attachments: currentPublication.attachments.filter((a) => a.id !== id)
     });
-    refreshPublications();
   };
+
+  const attachFile = async (fileName) => {
+    const attachment = {
+      id: Math.max(currentPublication.attachments.map((a) => a.id)) + 1,
+      userName: globalState.login,
+      fileName: fileName,
+      links: [{ rel: 'detach', href: 'createMode' }]
+    };
+
+    if (editMode) { // Just add to model
+      setCurrentPublication({
+        ...currentPublication,
+        attachments: [...currentPublication.attachments, attachment]
+      });
+      return;
+    }
+  };
+
+  const attachFilesInEditMode = (fileNames) => {
+    let freeId = Math.max(currentPublication.attachments.map((a) => a.id)) + 1;
+    const attachments = fileNames.map((f) => ({
+      id: freeId++,
+      userName: globalState.login,
+      fileName: f,
+      links: [{ rel: 'detach', href: 'createMode' }]
+    }));
+
+    setCurrentPublication({
+      ...currentPublication,
+      attachments: currentPublication.attachments.concat(attachments)
+    });
+  };
+
+  const attachFiles = async () => {
+    if (editMode) { // Just update model
+      attachFilesInEditMode(toAttachList);
+      setToAttachList([]);
+      return;
+    }
+  };
+
+
 
   const prepareAttachmentsOptions = () => {
     const attachedFileNames = currentPublication.attachments.map((a) => a.fileName);
@@ -175,18 +231,20 @@ const Publication = ({ createMode, publication, history, refreshPublications, gl
           </ul>
         </div>
 
-        {editMode &&
-          <div className="form-group ">
+        {editMode && (attachLink || createMode) &&
+          < div className="form-group ">
             <label htmlFor="attachments">Dołącz pliki</label>
             <div className="d-flex">
               <Select
                 isMulti
+                value={toAttachList.map((a) => ({ value: a, label: a }))}
                 name="attachments"
+                onChange={(selectedOptions) => setToAttachList(selectedOptions.map((o) => o.value))}
                 options={prepareAttachmentsOptions()}
                 className="basic-multi-select flex-grow-1 mr-2"
                 classNamePrefix="select"
               />
-              <button onClick={() => alert('Not implemented')} className="btn btn-success" type="button">
+              <button onClick={attachFiles} className="btn btn-success" type="button">
                 <i className="fas fa-link"></i>
                 <span className="d-none d-sm-inline ml-1">Dołącz</span>
               </button>
@@ -219,7 +277,7 @@ const Publication = ({ createMode, publication, history, refreshPublications, gl
           }
         </div>
       </form>
-    </section>
+    </section >
   );
 };
